refactor: split library filter entrance
